Fix Makefile Target for Tests
Previously make would skip making the tests even if the `tests/unittests`
file did not exist; becuase it was checking the directory existed
unchanged. This creates renames the old target to `tests/unittests` and
creates a new target `tests` that relies on it. This in effect fixes the
problem by forcing make to look for the unittests program